Refactor: move extensions sagas to separate files in extensions folder
