+++ src/modules/dashboard/sagas/extensions/colonyExtensionInstall.ts
@@ -1,0 +1,73 @@
+import { call, fork, put } from 'redux-saga/effects';
+import { ClientType, getExtensionHash } from '@colony/colony-js';
+
+import { Action, ActionTypes, AllActions } from '~redux/index';
+import {
+  NetworkExtensionVersionQuery,
+  NetworkExtensionVersionQueryVariables,
+  NetworkExtensionVersionDocument,
+} from '~data/index';
+import { ContextModule, TEMP_getContext } from '~context/index';
+import { putError, takeFrom } from '~utils/saga/effects';
+
+import {
+  createTransaction,
+  getTxChannel,
+  waitForTxResult,
+} from '../../../core/sagas';
+
+import { refreshExtension } from '../utils';
+
+export function* colonyExtensionInstall({
+  meta,
+  payload: { colonyAddress, extensionId },
+}: Action<ActionTypes.COLONY_EXTENSION_INSTALL>) {
+  const txChannel = yield call(getTxChannel, meta.id);
+  const apolloClient = TEMP_getContext(ContextModule.ApolloClient);
+
+  try {
+    /*
+     * Get the latest extension version that's deployed to the network
+     */
+    const {
+      data: { networkExtensionVersion },
+    } = yield apolloClient.query<
+      NetworkExtensionVersionQuery,
+      NetworkExtensionVersionQueryVariables
+    >({
+      query: NetworkExtensionVersionDocument,
+      variables: {
+        extensionId,
+      },
+      fetchPolicy: 'network-only',
+    });
+
+    yield fork(createTransaction, meta.id, {
+      context: ClientType.ColonyClient,
+      methodName: 'installExtension',
+      identifier: colonyAddress,
+      params: [getExtensionHash(extensionId), networkExtensionVersion],
+    });
+
+    yield takeFrom(txChannel, ActionTypes.TRANSACTION_CREATED);
+
+    yield put<AllActions>({
+      type: ActionTypes.COLONY_EXTENSION_INSTALL_SUCCESS,
+      payload: {},
+      meta,
+    });
+
+    yield waitForTxResult(txChannel);
+  } catch (error) {
+    return yield putError(
+      ActionTypes.COLONY_EXTENSION_INSTALL_ERROR,
+      error,
+      meta,
+    );
+  } finally {
+    yield call(refreshExtension, colonyAddress, extensionId);
+
+    txChannel.close();
+  }
+  return null;
+}
