+++ src/modules/dashboard/sagas/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './updateColonyDisplayCache';
 export { updateMotionValues } from './updateMotionValues';
+export { refreshExtension } from './refreshExtension';
